requirements.txt
streamlit
google-cloud-bigquery[pandas]
watchdog
google-cloud-bigquery-storage